requirements.txt
tensorflow>=2.13
tensorflow
numpy
matplotlib
fastapi
uvicorn
pillow
python-multipart